utils: move macros to utils::macros
